requirements.txt
pyodbc==5.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.54.1
graphviz==0.20.3
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
keeper-secrets-manager-core==16.6.6
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.9.2
numba==0.60.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pantab==5.0.0
pillow==10.4.0
pyarrow==17.0.0
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
shap==0.46.0
six==1.16.0
slicer==0.0.8
smartsheet-python-sdk==3.0.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
tableauhyperapi==0.0.20027
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2